packages/py-api/pyproject.toml
[tool.poetry]
name = "py-api"
version = "0.1.0"
description = ""
authors = ["Michael Bozhilov <michaelbozhilov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.101.1"
uvicorn = {extras = ["standard"], version = "^0.23.2"}
requests = "^2.31.0"
pymongo = "^4.5.0"
python-dotenv = "^1.0.0"

[tool.poetry.scripts]
start = "py_api.main:start"

[tool.poetry.group.dev.dependencies]
setuptools = "^68.1.2"
pre-commit = "^3.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
